Testing Google Antigravity cleanups on an old codebase.
