+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.8] - Unreleased
+### Added
+### Changed
+ - Automated refactoring using Google's AntiGravity IDE for testing.
+### Updated
+ - [Chia](https://github.com/Chia-Network/chia-blockchain/releases/tag/2.5.8) to v2.5.8 - misc improvements, see their release notes.
+
 ## [2.5.7] - 2025-11-21
 ### Added
 ### Changed
