+++ VERSION
@@ -1,1 +1,1 @@
-2.5.7
+2.5.8
