+++ docker/requirements.txt
@@ -1,27 +1,27 @@
-apscheduler
-click
-jinja2
-flask
-flask-smorest >= 0.42.1
-python-dotenv
-marshmallow
-sqlalchemy
-sqlalchemy-utils
-flask-sqlalchemy
-marshmallow-sqlalchemy
-marshmallow-toplevel
-gunicorn
-pytz
-wheel
-requests
-pexpect
-flask-migrate
-ruamel.yaml
-psutil
-pytest
-flask-babel
-beautifulsoup4
-geoip2
-gdown
-pendulum
-py-spy
+apscheduler==3.10.4
+click==8.1.7
+jinja2==3.1.2
+flask==3.0.0
+flask-smorest>=0.42.1
+python-dotenv==1.0.0
+marshmallow==3.20.1
+sqlalchemy==2.0.23
+sqlalchemy-utils==0.41.1
+flask-sqlalchemy==3.1.1
+marshmallow-sqlalchemy==0.29.0
+marshmallow-toplevel==0.1.2
+gunicorn==21.2.0
+pytz==2023.3.post1
+wheel==0.42.0
+requests==2.31.0
+pexpect==4.9.0
+flask-migrate==4.0.5
+ruamel.yaml==0.18.5
+psutil==5.9.6
+pytest==7.4.3
+flask-babel==4.0.0
+beautifulsoup4==4.12.2
+geoip2==4.7.0
+gdown==4.7.1
+pendulum==2.1.2
+py-spy==0.3.14
