+++ web/templates/alerts.html
@@ -117,7 +117,7 @@
             var width = 900;
             var top = (screen.height - height) / 2;
             var left = (screen.width - width) / 2;
-            window.open("{{ url_for('logs') }}?log=alerts&hostname=" + hostname + "&blockchain=" + blockchain, '{{_('Chiadog Log on ')}}' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+            window.open("{{ url_for('logs.index') }}?log=alerts&hostname=" + hostname + "&blockchain=" + blockchain, '{{_('Chiadog Log on ')}}' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
         }
         function MonitoringAction(hostname, service, action, status) {
             $("#" + hostname + "_btn").prop("disabled", true);
