+++ web/templates/blockchains.html
@@ -117,7 +117,7 @@
             <tr>
                 <td>
                   {% if blockchain.blockchain in fullnodes %}
-                  <a class="text-white" href="{{ url_for('worker_route') }}?hostname={{ fullnodes[blockchain.blockchain].hostname }}&blockchain={{ blockchain.blockchain }}">{{blockchain.blockchain}}</a>
+                  <a class="text-white" href="{{ url_for('workers.worker_route') }}?hostname={{ fullnodes[blockchain.blockchain].hostname }}&blockchain={{ blockchain.blockchain }}">{{blockchain.blockchain}}</a>
                   {% else %}
                   {{blockchain.blockchain}}
                   {% endif %}
@@ -211,7 +211,7 @@ function ViewLogs(hostname, blockchain) {
     var width = 900;
     var top = (screen.height - height) / 2;
     var left = (screen.width - width) / 2;
-    window.open("{{ url_for('logs') }}?log=farming" + "&hostname=" + hostname + "&blockchain=" + blockchain, '{{_('Fullnode Log for ')}}' + blockchain + ' - ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+    window.open("{{ url_for('logs.index') }}?log=farming" + "&hostname=" + hostname + "&blockchain=" + blockchain, '{{_('Fullnode Log for ')}}' + blockchain + ' - ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
 }
 $(document).ready(function () {
     $("#btnSave").click(function () {
