+++ web/templates/farming/plots.html
@@ -202,13 +202,13 @@
     <div class="container-fluid">
         <ul class="nav nav-tabs bg-dark">
             <li class="nav-item">
-                <a class="nav-link active" aria-current="page" href="{{ url_for('farming_plots') }}">{{_('Plots')}}</a>
+                <a class="nav-link active" aria-current="page" href="{{ url_for('farming.plots') }}">{{_('Plots')}}</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="{{ url_for('farming_workers') }}">{{_('Workers')}}</a>
+                <a class="nav-link" href="{{ url_for('farming.workers') }}">{{_('Workers')}}</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="{{ url_for('farming_warnings') }}">{{_('Warnings')}}</a>
+                <a class="nav-link" href="{{ url_for('farming.warnings_page') }}">{{_('Warnings')}}</a>
             </li>
         </ul>
         <form class="d-flex" method="POST" style="padding-bottom:5px" id="main-form">
@@ -295,13 +295,13 @@
             var width = 900;
             var top = (screen.height - height) / 2;
             var left = (screen.width - width) / 2;
-            window.open("{{ url_for('logs') }}?log=farming" + "&hostname=" + hostname + "&blockchain=" + blockchain, '{{_('Farming Log for ')}}' + blockchain + ' - ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+            window.open("{{ url_for('logs.index') }}?log=farming" + "&hostname=" + hostname + "&blockchain=" + blockchain, '{{_('Farming Log for ')}}' + blockchain + ' - ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
         }
         function PlotAnalyze(plot) {
             var output = document.getElementById('plotAnalyzeContent');
             output.textContent = ""; // Clear any old analysis before requesting
             var xhr = new XMLHttpRequest();
-            xhr.open('GET', "{{ url_for('farming_plots') }}?analyze=" + plot);
+            xhr.open('GET', "{{ url_for('farming.plots') }}?analyze=" + plot);
             xhr.send();
             xhr.onload = function () {
                 output.textContent = xhr.responseText;
@@ -313,7 +313,7 @@
             var plotCheckId = document.getElementById('plotCheckId');
             plotCheckId.value = plot;
             var xhr = new XMLHttpRequest();
-            xhr.open('GET', "{{ url_for('farming_plots') }}?check=" + plot);
+            xhr.open('GET', "{{ url_for('farming.plots') }}?check=" + plot);
             xhr.send();
             xhr.onload = function () {
                 output.textContent = xhr.responseText;
@@ -331,7 +331,7 @@
             $("#btnRecheckPlot").click(function () {
                 var plotCheckId = document.getElementById('plotCheckId');
                 var xhr = new XMLHttpRequest();
-                xhr.open('GET', "{{ url_for('farming_plots') }}?check=" + plotCheckId.value + "&force_recheck=true");
+                xhr.open('GET', "{{ url_for('farming.plots') }}?check=" + plotCheckId.value + "&force_recheck=true");
                 xhr.send();
                 alert("Plot check initiated. Please refresh in a few minutes.\n\nNote that all plots will be checked slowly in the background so you don't need to check each one individually.  Please be patient.");
             });
@@ -339,7 +339,7 @@
                 "stateSave": true,
                 "pageLength": 25,
                 "serverSide": true,
-                "ajax": "{{ url_for('farming_data') }}",
+                "ajax": "{{ url_for('farming.data') }}",
                 "order": [[6, "desc"]],
                 "columnDefs": [
                     {
