+++ web/templates/farming/warnings.html
@@ -22,13 +22,13 @@
     <div class="container-fluid">
         <ul class="nav nav-tabs bg-dark">
             <li class="nav-item">
-                <a class="nav-link" href="{{ url_for('farming_plots') }}">{{_('Plots')}}</a>
+                <a class="nav-link" href="{{ url_for('farming.plots') }}">{{_('Plots')}}</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="{{ url_for('farming_workers') }}">{{_('Workers')}}</a>
+                <a class="nav-link" href="{{ url_for('farming.workers') }}">{{_('Workers')}}</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link active" aria-current="page" href="{{ url_for('farming_warnings') }}">{{_('Warnings')}}</a>
+                <a class="nav-link active" aria-current="page" href="{{ url_for('farming.warnings_page') }}">{{_('Warnings')}}</a>
             </li>
         </ul>
     </div>
