+++ web/templates/index.html
@@ -29,7 +29,7 @@
             <div class="h-30 p-2 text-white bg-dark rounded-3">
               <h2 class="text-center" style="margin-bottom: 0px">{{ blockchain|capitalize }} - {{_('Expected Time to Win')}}:
                 <span class="text-success">{{ farms[blockchain].expected_time_to_win }}</span>
-                <a href="{{ url_for('index') }}?selected_blockchain={{blockchain}}" title="{{_('Pin Current Blockchain')}}">
+                <a href="{{ url_for('index.index') }}?selected_blockchain={{blockchain}}" title="{{_('Pin Current Blockchain')}}">
                   <i class="fs-4 bi-pin"></i>
                 </a>
               </h2>
@@ -196,7 +196,7 @@
       var width = 900;
       var top = (screen.height - height) / 2;
       var left = (screen.width - width) / 2;
-      window.open("{{ url_for('chart') }}?type=" + chart_type + "&blockchain=" + blockchain, blockchain + ' - ' + chart_type, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+      window.open("{{ url_for('index.chart') }}?type=" + chart_type + "&blockchain=" + blockchain, blockchain + ' - ' + chart_type, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
     }
   </script>
   {% endif %}
