+++ web/templates/landing.html
@@ -38,7 +38,7 @@
       margin: 10px;
     }
   </style>
-  <meta http-equiv="refresh" content="3;url={{ url_for('index') }}">
+  <meta http-equiv="refresh" content="3;url={{ url_for('index.index') }}">
 </head>
 
 <body>
