+++ web/templates/logs.html
@@ -38,7 +38,7 @@
             if (document.getElementById('autoscroll').checked == 1) {
                 var output = document.getElementById('output');
                 var xhr = new XMLHttpRequest();
-                xhr.open('GET', "{{ url_for('logfile') }}?hostname=" + get('hostname') + "&log=" + get('log') + "&log_id=" + get('log_id') + "&blockchain=" + get('blockchain'));
+                xhr.open('GET', "{{ url_for('logs.file') }}?hostname=" + get('hostname') + "&log=" + get('log') + "&log_id=" + get('log_id') + "&blockchain=" + get('blockchain'));
                 xhr.send();
                 xhr.onload = function () {
                     output.textContent = xhr.responseText;
