+++ web/templates/plotting/jobs.html
@@ -103,13 +103,13 @@
     <div class="container-fluid">
         <ul class="nav nav-tabs bg-dark">
             <li class="nav-item">
-                <a class="nav-link active" aria-current="page" href="{{ url_for('plotting_jobs') }}">{{_('Jobs')}}</a>
+                <a class="nav-link active" aria-current="page" href="{{ url_for('plotting.jobs') }}">{{_('Jobs')}}</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="{{ url_for('plotting_transfers') }}">{{_('Transfers')}}</a>
+                <a class="nav-link" href="{{ url_for('plotting.transfers') }}">{{_('Transfers')}}</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="{{ url_for('plotting_workers') }}">{{_('Workers')}}</a>
+                <a class="nav-link" href="{{ url_for('plotting.workers') }}">{{_('Workers')}}</a>
             </li>
         </ul>
         <form method="POST" id="plotman-form" style="padding-bottom: 10px;">
@@ -349,9 +349,9 @@
             var top = (screen.height - height) / 2;
             var left = (screen.width - width) / 2;
             if (typeof plot_id !== 'undefined') {
-                window.open("{{ url_for('logs') }}?log=" + type + "&hostname=" + hostname + "&blockchain=" + blockchain + "&log_id=" + plot_id, 'Plotman Log ' + plot_id + ' on ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+                window.open("{{ url_for('logs.index') }}?log=" + type + "&hostname=" + hostname + "&blockchain=" + blockchain + "&log_id=" + plot_id, 'Plotman Log ' + plot_id + ' on ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
             } else {
-                window.open("{{ url_for('logs') }}?log=" + type + "&hostname=" + hostname + "&blockchain=" + blockchain, 'Plotman Log on ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+                window.open("{{ url_for('logs.index') }}?log=" + type + "&hostname=" + hostname + "&blockchain=" + blockchain, 'Plotman Log on ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
             }
         }
         function modal() {
@@ -359,7 +359,7 @@
         }
         function load_dirs(worker, blockchain) {
             jqxhr = $.getJSON({
-                url: "{{ url_for('views_settings_config') }}?type=plotting_dirs&worker=" + worker + "&blockchain=" + blockchain,
+                url: "{{ url_for('settings.config') }}?type=plotting_dirs&worker=" + worker + "&blockchain=" + blockchain,
             })
                 .done(function (data) {
                     if ('errors' in data) {
