+++ web/templates/settings/alerts.html
@@ -108,7 +108,7 @@
                 $("#btnDownload").prop('disabled', true);
                 $.ajax({
                     type: "GET",
-                    url: "{{ url_for('views_settings_config') }}?type=alerts&worker=" + worker + "&blockchain=" + blockchain,
+                    url: "{{ url_for('settings.config') }}?type=alerts&worker=" + worker + "&blockchain=" + blockchain,
                 })
                     .done(function (data) {
                         $("#config").val(data)
