+++ web/templates/settings/farming.html
@@ -79,7 +79,7 @@
                 $("#btnDownload").prop('disabled', true);
                 $.ajax({
                     type: "GET",
-                    url: "{{ url_for('views_settings_config') }}?type=farming&worker=" + worker + "&blockchain=" + blockchain,
+                    url: "{{ url_for('settings.config') }}?type=farming&worker=" + worker + "&blockchain=" + blockchain,
                 })
                     .done(function (data) {
                         var target_addresses = new Set();
