+++ web/templates/settings/pools.html
@@ -244,7 +244,7 @@
             var width = 900;
             var top = (screen.height - height) / 2;
             var left = (screen.width - width) / 2;
-            window.open("{{ url_for('logs') }}?log=pooling&hostname=" + hostname + "&blockchain=" + blockchain, '{{_('PlotNFT Log on ')}}' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+            window.open("{{ url_for('logs.index') }}?log=pooling&hostname=" + hostname + "&blockchain=" + blockchain, '{{_('PlotNFT Log on ')}}' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
         }
     </script>
     {% endblock %}
