+++ web/templates/wallet.html
@@ -359,11 +359,11 @@
             var width = 900;
             var top = (screen.height - height) / 2;
             var left = (screen.width - width) / 2;
-            window.open("{{ url_for('logs') }}?log=rewards" + "&hostname=" + hostname + "&blockchain=" + blockchain, '{{_('Reward Recovery Log for ')}}' + blockchain + ' - ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+            window.open("{{ url_for('logs.index') }}?log=rewards" + "&hostname=" + hostname + "&blockchain=" + blockchain, '{{_('Reward Recovery Log for ')}}' + blockchain + ' - ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
         }
         function load_rewards() {
             jqxhr = $.getJSON({
-                url: "{{ url_for('wallet') }}?rewards=get",
+                url: "{{ url_for('wallet.index') }}?rewards=get",
             })
                 .done(function (data) {
                     if (Object.entries(data).length > 0) {
@@ -392,7 +392,7 @@
           var width = 900;
           var top = (screen.height - height) / 2;
           var left = (screen.width - width) / 2;
-          window.open("{{ url_for('transactions') }}?blockchain=" + blockchain + "&selected_wallet_id=" + wallet_id, blockchain + '{{_(' Transactions')}}: {{_('Wallet ID')}}#'+wallet_id, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+          window.open("{{ url_for('transactions.index') }}?blockchain=" + blockchain + "&selected_wallet_id=" + wallet_id, blockchain + '{{_(' Transactions')}}: {{_('Wallet ID')}}#'+wallet_id, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
         }
         function PopupChart(chart_type, blockchain) {
           var d = new Date();
@@ -400,7 +400,7 @@
           var width = 900;
           var top = (screen.height - height) / 2;
           var left = (screen.width - width) / 2;
-          window.open("{{ url_for('chart') }}?type=" + chart_type + "&blockchain=" + blockchain, blockchain + ' - ' + chart_type, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+          window.open("{{ url_for('index.chart') }}?type=" + chart_type + "&blockchain=" + blockchain, blockchain + ' - ' + chart_type, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
         }
         const COLORS = [
           '#3aac59',
