+++ web/templates/workers.html
@@ -134,7 +134,7 @@
         var width = 900;
         var top = (screen.height - height) / 2;
         var left = (screen.width - width) / 2;
-        window.open("{{ url_for('chart') }}?type=" + chart_type + "&hostname=" + hostname + "&blockchain=" + blockchain, blockchain + ' - ' + chart_type, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+        window.open("{{ url_for('index.chart') }}?type=" + chart_type + "&hostname=" + hostname + "&blockchain=" + blockchain, blockchain + ' - ' + chart_type, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
     }
     function WorkerLaunch(hostname) {
         var d = new Date();
@@ -142,7 +142,7 @@
         var width = 900;
         var top = (screen.height - height) / 2;
         var left = (screen.width - width) / 2;
-        window.open("{{ url_for('worker_launch') }}", 'Worker Launch Config', 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+        window.open("{{ url_for('workers.launch') }}", 'Worker Launch Config', 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
     }
     function MachinarisLog(type, hostname, blockchain) {
         var d = new Date();
@@ -150,7 +150,7 @@
         var width = 900;
         var top = (screen.height - height) / 2;
         var left = (screen.width - width) / 2;
-        window.open("{{ url_for('logs') }}?log=" + type + "&hostname=" + hostname + "&blockchain=" + blockchain, 'Machinaris ' + type + '.log on ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
+        window.open("{{ url_for('logs.index') }}?log=" + type + "&hostname=" + hostname + "&blockchain=" + blockchain, 'Machinaris ' + type + '.log on ' + hostname, 'resizeable=yes,scrollbars=yes,height=' + height + ',width=' + width + ',top=' + top + ',left=' + left).focus();
     }
     $(document).ready(function () {
       $('#data').DataTable({
@@ -172,7 +172,7 @@
             { 
                 targets: [2], "orderable": true, 
                 "render": function ( data, type, row, meta ) {
-                    var url = "{{ url_for('worker_route') }}";
+                    var url = "{{ url_for('workers.worker_route') }}";
                     arr = row[0].split('|');
                     hostname = arr[0];
                     blockchain = arr[1];
