+++ web/utils.py
@@ -76,3 +76,34 @@ def convert_chia_ip_address(chia_ip_address):
     if chia_ip_address in ['127.0.0.1']:
         return get_hostname()
     return chia_ip_address  # TODO Map duplicated IPs from docker internals...
+
+def get_lang(request):
+    try:
+        accept = request.headers['Accept-Language']
+        match = request.accept_languages.best_match(app.config['LANGUAGES'])
+        # Workaround for dumb babel match method suggesting 'en' for 'nl' instead of 'nl_NL'
+        if match == 'en' and not accept.startswith('en'):
+            first_accept = accept.split(',')[0]  # Like 'nl'
+            alternative = "{0}_{1}".format(first_accept, first_accept.upper())
+            if alternative in app.config['LANGUAGES']:
+                app.logger.debug("LOCALE: Accept-Language: {0}  ---->  using locale: {1}".format(accept, alternative))
+                return alternative
+        if match:
+            app.logger.debug("LOCALE: Accept-Language: {0}  ---->  matched locale: {1}".format(accept, match))
+            return match
+        app.logger.debug("LOCALE: Accept-Language: {0} returned no match so defaulting to 'en'.".format(accept))
+        return "en"
+    except:
+        app.logger.debug("LOCALE: Request had no Accept-Language header, returning default locale of 'en'")
+        return "en"
+
+def find_selected_worker(hosts, hostname, blockchain= None):
+    if len(hosts) == 0:
+        return None
+    if not blockchain:
+        pass # hosts[0].workers[0] # Original code had this no-op
+    for host in hosts:
+        for worker in host.workers:
+            if worker['hostname'] == hostname and worker['blockchain'] == blockchain:
+                return worker
+    return hosts[0].workers[0]
